@@ -2,8 +2,10 @@ package api
 
 import (
 	"encoding/json"
-	"log"
+	"log/slog"
 	"net/http"
+
+	"github.com/Rainminds/gantral/core/errors"
 )
 
 // handleCreateInstance handles the creation of a new execution instance.
@@ -11,7 +13,7 @@ func (s *Server) handleCreateInstance(w http.ResponseWriter, r *http.Request) {
 	// 1. Decode generic JSON body (for now just validating JSON structure)
 	var body map[string]interface{}
 	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
-		http.Error(w, "Invalid request body", http.StatusBadRequest)
+		s.writeError(w, errors.Wrap(errors.ErrInvalidInput, "invalid json body"))
 		return
 	}
 
@@ -27,7 +29,7 @@ func (s *Server) handleCreateInstance(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(http.StatusCreated)
 	if err := json.NewEncoder(w).Encode(response); err != nil {
-		log.Printf("Failed to encode response: %v", err)
+		slog.Error("failed to encode response", "error", err)
 	}
 }
 
@@ -36,12 +38,18 @@ func (s *Server) handleGetInstance(w http.ResponseWriter, r *http.Request) {
 	// 1. Extract ID using Go 1.22 PathValue
 	id := r.PathValue("id")
 	if id == "" {
-		http.Error(w, "Missing instance ID", http.StatusBadRequest)
+		s.writeError(w, errors.Wrap(errors.ErrInvalidInput, "missing instance id"))
 		return
 	}
 
 	// 2. Call Engine (Stubbed)
 	// In future: instance, err := s.engine.GetInstance(id)
+
+	// Stub logic for 404
+	if id == "notfound" {
+		s.writeError(w, errors.ErrNotFound)
+		return
+	}
 
 	// 3. Response
 	response := map[string]interface{}{
@@ -53,6 +61,39 @@ func (s *Server) handleGetInstance(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(http.StatusOK)
 	if err := json.NewEncoder(w).Encode(response); err != nil {
-		log.Printf("Failed to encode response: %v", err)
+		slog.Error("failed to encode response", "error", err)
 	}
+}
+
+// handleHealthz returns a simple 200 OK status.
+func (s *Server) handleHealthz(w http.ResponseWriter, r *http.Request) {
+	w.WriteHeader(http.StatusOK)
+	w.Write([]byte("ok"))
+}
+
+// writeError maps domain errors to HTTP status codes and writes a JSON error response.
+func (s *Server) writeError(w http.ResponseWriter, err error) {
+	var status int
+	switch {
+	case errors.Is(err, errors.ErrInvalidInput):
+		status = http.StatusBadRequest
+	case errors.Is(err, errors.ErrNotFound):
+		status = http.StatusNotFound
+	case errors.Is(err, errors.ErrConflict):
+		status = http.StatusConflict
+	case errors.Is(err, errors.ErrUnauthorized):
+		status = http.StatusUnauthorized
+	default:
+		status = http.StatusInternalServerError
+	}
+
+	if status == http.StatusInternalServerError {
+		slog.Error("internal server error", "error", err)
+	}
+
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(status)
+	json.NewEncoder(w).Encode(map[string]string{
+		"error": err.Error(),
+	})
 }
